examen/pyproject.toml
[project]
name = "examen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.8",
    "mpi4py>=4.1.0",
    "numba>=0.64.0",
    "pandas>=3.0.1",
    "pyopengl>=3.1.10",
    "pysdl2>=0.9.17",
]
